[ScSynth] - Store all args and send on creation
It adds storing other types of args other than float and int values, and also sets them in the correct way inside the "/n_new" message instead of bundling messages.
To send vector we need the "Array Begin" and "Array End" osc typetags, but as ofxOsc does not implement them we hacked our ofxOscSenderReceiver custom implementation to send this messages  when the '[' and ']' characters are received, as they are the tag aliases for this typetags, and as in supercollider there are no osc char types should be safe to use.